proper credit deduction

+++ app/models/tenant.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, String, DateTime, Integer
+from sqlalchemy import Column, String, DateTime, Integer, Numeric
 from sqlalchemy.dialects.postgresql import UUID
 from sqlalchemy.orm import relationship
 from sqlalchemy.sql import func
@@ -13,7 +13,7 @@ class Tenant(Base):
     stripe_customer_id = Column(String, nullable=True, index=True)
     stripe_subscription_id = Column(String, nullable=True, index=True)
     created_at = Column(DateTime(timezone=True), server_default=func.now(), nullable=False)
-    credits = Column(Integer, default=0, nullable=False)  # New field for credit system, now NOT NULL
+    credits = Column(Numeric(10, 4), default=0, nullable=False)  # Float credits with 4 decimal precision
     
     # Relationships
     users = relationship("User", secondary="user_tenant_association", back_populates="tenants")
